Split test cases in QueryEngineSpec into HAPlannerSpec and SingleClusterPlannerSpec
