+++ coordinator/src/main/scala/filodb.coordinator/QueryActor.scala
@@ -12,7 +12,7 @@ import monix.execution.Scheduler
 import net.ceedubs.ficus.Ficus._
 import net.ceedubs.ficus.readers.ValueReader
 
-import filodb.coordinator.queryplanner.{CompositePlanner, EmptyFailureProvider}
+import filodb.coordinator.queryplanner.SingleClusterPlanner
 import filodb.core._
 import filodb.core.memstore.{FiloSchedulers, MemStore, TermInfo}
 import filodb.core.metadata.Schemas
@@ -79,8 +79,7 @@ final class QueryActor(memStore: MemStore,
   val functionalSpreadProvider = FunctionalSpreadProvider(spreadFunc)
 
   logger.info(s"Starting QueryActor and QueryEngine for ds=$dsRef schemas=$schemas")
-  val queryEngine2 = new CompositePlanner(dsRef, schemas, shardMapFunc,
-    EmptyFailureProvider, functionalSpreadProvider)
+  val queryPlanner = new SingleClusterPlanner(dsRef, schemas, shardMapFunc, functionalSpreadProvider)
   val queryConfig = new QueryConfig(config.getConfig("filodb.query"))
   val numSchedThreads = Math.ceil(config.getDouble("filodb.query.threads-factor") * sys.runtime.availableProcessors)
   val queryScheduler = Scheduler.fixedPool(s"$QuerySchedName-$dsRef", numSchedThreads.toInt)
@@ -124,7 +123,7 @@ final class QueryActor(memStore: MemStore,
     // This is for CLI use only. Always prefer clients to materialize logical plan
     lpRequests.increment
     try {
-      val execPlan = queryEngine2.materialize(q.logicalPlan, q.qContext)
+      val execPlan = queryPlanner.materialize(q.logicalPlan, q.qContext)
       self forward execPlan
     } catch {
       case NonFatal(ex) =>
@@ -136,7 +135,7 @@ final class QueryActor(memStore: MemStore,
 
   private def processExplainPlanQuery(q: ExplainPlan2Query, replyTo: ActorRef) = {
     try {
-      val execPlan = queryEngine2.materialize(q.logicalPlan, q.qContext)
+      val execPlan = queryPlanner.materialize(q.logicalPlan, q.qContext)
       replyTo ! execPlan
     } catch {
       case NonFatal(ex) =>
