+++ coordinator/src/test/scala/filodb.coordinator/client/SerializationSpec.scala
@@ -6,7 +6,7 @@ import akka.testkit.TestProbe
 import org.scalatest.concurrent.ScalaFutures
 
 import filodb.coordinator.{ActorSpecConfig, ActorTest, ShardMapper}
-import filodb.coordinator.queryplanner.{EmptyFailureProvider, CompositePlanner}
+import filodb.coordinator.queryplanner.SingleClusterPlanner
 import filodb.core.{query, MachineMetricsData, SpreadChange}
 import filodb.core.binaryrecord2.BinaryRecordRowReader
 import filodb.core.metadata.{Dataset, Schemas}
@@ -16,7 +16,7 @@ import filodb.memory.format.{RowReader, SeqRowReader, UTF8MapIteratorRowReader, 
 import filodb.prometheus.ast.TimeStepParams
 import filodb.prometheus.parse.Parser
 import filodb.query.{QueryResult => QueryResult2, _}
-import filodb.query.exec.{PartKeysDistConcatExec}
+import filodb.query.exec.PartKeysDistConcatExec
 
 object SerializationSpecConfig extends ActorSpecConfig {
   override val defaultConfig = """
@@ -172,7 +172,7 @@ class SerializationSpec extends ActorTest(SerializationSpecConfig.getNewSystem) 
     val mapper = new ShardMapper(1)
     mapper.registerNode(Seq(0), node0)
     def mapperRef: ShardMapper = mapper
-    val engine = new CompositePlanner(dataset.ref, Schemas.global, mapperRef, EmptyFailureProvider)
+    val engine = new SingleClusterPlanner(dataset.ref, Schemas.global, mapperRef)
     val f1 = Seq(ColumnFilter("__name__", Filter.Equals("http_request_duration_seconds_bucket")),
       ColumnFilter("job", Filter.Equals("myService")),
       ColumnFilter("le", Filter.Equals("0.3")),
@@ -211,7 +211,7 @@ class SerializationSpec extends ActorTest(SerializationSpecConfig.getNewSystem) 
     val to = System.currentTimeMillis() / 1000
     val from = to - 50
     val qParams = TimeStepParams(from, 10, to)
-    val engine = new CompositePlanner(dataset.ref, Schemas.global, mapperRef, EmptyFailureProvider)
+    val engine = new SingleClusterPlanner(dataset.ref, Schemas.global, mapperRef)
 
     val logicalPlan1 = Parser.queryRangeToLogicalPlan(
       s"""sum(rate(http_request_duration_seconds_bucket{job="prometheus",$shardKeyStr}[20s])) by (handler)""",
@@ -240,7 +240,7 @@ class SerializationSpec extends ActorTest(SerializationSpecConfig.getNewSystem) 
     val to = System.currentTimeMillis() / 1000
     val from = to - 50
     val qParams = TimeStepParams(from, 10, to)
-    val engine = new CompositePlanner(dataset.ref, Schemas.global, mapperRef, EmptyFailureProvider)
+    val engine = new SingleClusterPlanner(dataset.ref, Schemas.global, mapperRef)
 
     // with column filters having shardcolumns
     val logicalPlan1 = Parser.metadataQueryToLogicalPlan(
